feat!: remove original otlp studio custom format exporter
since the trace export format is an implementation detail of our
stack, we do not need to offer both options to the user
