+++ pharia_skill/studio/__init__.py
@@ -1,14 +1,9 @@
 from .client import StudioClient
-from .exporter import SpanClient, StudioExporter
-from .otlp_exporter import StudioOTLPSpanExporter
+from .exporter import StudioExporter
 from .processor import StudioSpanProcessor
-from .span import StudioSpan
 
 __all__ = [
     "StudioClient",
     "StudioExporter",
-    "StudioOTLPSpanExporter",
-    "StudioSpan",
-    "SpanClient",
     "StudioSpanProcessor",
 ]
