+++ pharia_skill/studio/client.py
@@ -1,5 +1,4 @@
 import os
-from collections.abc import Sequence
 from typing import Optional, cast
 from urllib.parse import urljoin
 
@@ -8,16 +7,13 @@ from dotenv import load_dotenv
 from pydantic import BaseModel
 from requests.exceptions import ConnectionError, MissingSchema
 
-from pharia_skill.studio.exporter import SpanClient
-from pharia_skill.studio.span import StudioSpan, StudioSpanList
-
 
 class StudioProject(BaseModel):
     name: str
     description: Optional[str]
 
 
-class StudioClient(SpanClient):
+class StudioClient:
     """Client for communicating with Pharia Studio.
 
     The Studio instance is determined by the environment variable `PHARIA_STUDIO_ADDRESS`.
@@ -26,10 +22,7 @@ class StudioClient(SpanClient):
       project_id (int, required): The unique identifier of the project currently in use.
     """
 
-    def __init__(
-        self,
-        project_name: str,
-    ) -> None:
+    def __init__(self, project_name: str) -> None:
         """Initializes the client.
 
         Runs a health check to check for a valid url of the Studio connection.
@@ -140,38 +133,6 @@ class StudioClient(SpanClient):
             case _:
                 response.raise_for_status()
         return cast(str, response.json())
-
-    def submit_spans(self, spans: Sequence[StudioSpan]) -> None:
-        """Sends the provided spans to Studio as a singular trace.
-
-        The method fails if the span list is empty, has already been created or if
-        spans belong to multiple traces.
-
-        Args:
-            spans (Sequence[StudioSpan], required): Spans to create the trace from. Created by exporting from a :class:`Tracer`.
-        """
-        if len(spans) == 0:
-            raise ValueError("Tried to upload an empty trace")
-        self._upload_trace(StudioSpanList(spans))
-
-    def _upload_trace(self, trace: StudioSpanList) -> None:
-        url = urljoin(self.url, f"/api/projects/{self.project_id}/traces")
-        response = requests.post(
-            url,
-            data=trace.model_dump_json(),
-            headers=self._headers,
-        )
-        match response.status_code:
-            case 409:
-                raise ValueError(
-                    f"Trace with id {trace.root[0].context.trace_id} already exists."
-                )
-            case 422:
-                raise ValueError(
-                    f"Uploading the trace failed with 422. Response: {response.json()}"
-                )
-            case _:
-                response.raise_for_status()
 
     def list_traces(self) -> list[str]:
         url = urljoin(self.url, f"/api/projects/{self.project_id}/traces_v2")
